+++ src/Settings.php
@@ -493,9 +493,6 @@ class Settings {
         // get main tab from request.
         $current_tab = filter_input( INPUT_GET, 'tab', FILTER_SANITIZE_FULL_SPECIAL_CHARS );
 
-        // get sub tab from request.
-        $current_sub_tab = filter_input( INPUT_GET, 'sub_tab', FILTER_SANITIZE_FULL_SPECIAL_CHARS );
-
         // sort the tabs.
         add_filter( $this->get_slug() . '_settings_tabs' , array( $this, 'sort' ), PHP_INT_MAX );
 
@@ -1296,7 +1293,7 @@ class Settings {
         wp_enqueue_script(
             $this->get_slug() . '-settings',
             $this->get_url() . 'Files/js.js',
-            array( 'jquery', 'easy-dialog-for-wordpress' ),
+            array( 'jquery', 'jquery-ui-sortable', 'jquery-ui-draggable', 'jquery-ui-droppable' ),
             (string) filemtime( $this->get_path() . 'Files/js.js' ),
             true
         );
@@ -1326,7 +1323,8 @@ class Settings {
                 'rest_nonce' => wp_create_nonce( 'wp_rest' ),
                 'title_add_image' => __( 'Add file' ),
                 'button_add_image' => __( 'Choose file' ),
-                'lbl_upload_image' => __( 'Upload or choose image' )
+                'lbl_upload_image' => __( 'Upload or choose image' ),
+                'label_sortable_title' => __( 'Hold to drag & drop' )
             )
         );
     }